requirements.txt
fastapi==0.109.1
pydantic==2.11.4
pydantic-settings==2.9.1
starlette==0.35.1

uvicorn[standard]==0.22.0
sqlalchemy==1.4.47

requests==2.32.0
httpx==0.27.0
cryptography==44.0.1
pyodbc~=5.2.0
python-dotenv~=1.1.0

nltk~=3.9.1               # используется в sentiment-анализе
sentiment~=0.0.1          # если кастомная библиотека есть

# Strawberry вместо Ariadne и gql
strawberry-graphql[fastapi]==0.217.1  # последняя стабильная версия
# Убираем ariadne и gql, не нужны для Strawberry

# поддержка C и сетевых библиотек
cffi~=1.17.1
pycparser~=2.22

# доп. утилиты и совместимость
h11~=0.16.0
httpcore~=1.0.9
sniffio~=1.3.1
anyio~=4.9.0
click~=8.2.0
typing_extensions~=4.13.2
exceptiongroup~=1.3.0
annotated-types~=0.7.0
charset-normalizer~=3.4.2
certifi~=2025.4.26
urllib3~=2.4.0
idna~=3.10
httptools~=0.6.4
websockets~=15.0.1
watchfiles~=1.0.5
greenlet~=3.2.2

# build инструменты
pip~=25.1.1
setuptools~=70.0.0
wheel~=0.41.2
PyYAML~=6.0.2

auth~=0.5.3
gunicorn~=23.0.0
packaging~=25.0
Werkzeug~=3.1.3
passlib~=1.7.4
python-jose~=3.4.0
attrs~=25.3.0
python-multipart~=0.0.20
async-timeout~=5.0.1
